test: 깨지는 테스트 코드 수정
