+++ tomcat/src/test/java/org/apache/coyote/http11/handler/StaticResourceHttpHandlerTest.java
@@ -3,7 +3,6 @@ package org.apache.coyote.http11.handler;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.File;
-import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
 import org.apache.coyote.http11.message.request.HttpMethod;
@@ -17,14 +16,15 @@ class StaticResourceHttpHandlerTest {
 
     @Test
     @DisplayName("정적 리소스 요청을 받으면 해당 리소스를 반환한다.")
-    void handleTest() throws IOException {
+    void handleTest() throws Exception {
         // given
         StaticResourceHttpHandler staticResourceHttpHandler = new StaticResourceHttpHandler();
         HttpRequest request = new HttpRequest(HttpMethod.GET, new HttpUrl("/index.html"));
         final URL resource = getClass().getClassLoader().getResource("static/index.html");
 
         // when
-        HttpResponse response = staticResourceHttpHandler.handle(request);
+        HttpResponse response = new HttpResponse();
+        staticResourceHttpHandler.service(request, response);
         byte[] expected = Files.readAllBytes(new File(resource.getFile()).toPath());
 
         // then
