+++ tomcat/src/test/java/org/apache/coyote/http11/handler/WelcomePageControllerTest.java
@@ -2,7 +2,7 @@ package org.apache.coyote.http11.handler;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.io.IOException;
+import com.techcourse.controller.WelcomePageController;
 import org.apache.coyote.http11.message.request.HttpMethod;
 import org.apache.coyote.http11.message.request.HttpRequest;
 import org.apache.coyote.http11.message.request.HttpUrl;
@@ -10,19 +10,20 @@ import org.apache.coyote.http11.message.response.HttpResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class StringHttpHandlerTest {
+class WelcomePageControllerTest {
 
     @Test
     @DisplayName("StringHttpHandler는 요청에 대한 응답으로 정적 문자열을 반환한다.")
-    void handleTest() throws IOException {
+    void handleTest() throws Exception {
         // given
-        StringHttpHandler stringHttpHandler = new StringHttpHandler("Hello, World!");
+        WelcomePageController stringHttpHandler = new WelcomePageController();
         HttpRequest request = new HttpRequest(HttpMethod.GET, new HttpUrl("/"));
 
         // when
-        HttpResponse response = stringHttpHandler.handle(request);
+        HttpResponse response = new HttpResponse();
+        stringHttpHandler.service(request, response);
 
         // then
-        assertEquals("Hello, World!", new String(response.getBody()));
+        assertEquals("Hello world!", new String(response.getBody()));
     }
 }
